Added a test for static if...else if mistakes.
If you write your code as:
static if (foo)
  doFoo();
else if (bar)
  doBar();

This check will catch it. If you write it as:
static if (foo)
  doFoo();
else {
  if (bar)
    doBar();
}

This isn't a bug and we won't warn about it.

If you write your code in an intermediate style, we'll still warn about it:
static if (foo)
  doFoo();
else
  if (bar)
    doBar();

Which probably isn't optimal, but what the hey.