add test for ignored & templated unittests
in DMD it seems to be a syntax error to have unittests in functions or
other unittests, but in libdparse it's allowed, so this dscanner check
adds a warning/error for that.

In the same check it implements a warning for unittests which are used
in templates. However as unittests in templates could be there just for
documentation, no warning is issued if annotated with a doc comment.

Mixin templates are not regarded as templates in this check as they
might be used for automatically generating unittests outside templates.